+++ tests/testthat/_snaps/spin/spin_preamble-custom-preamble.R
@@ -1,0 +1,9 @@
+#' ---
+#' title: My Report
+#' author: John Doe
+#' format: html
+#' ---
+#' 
+
+library(ggplot2)
+plot(1:10)